app/requirements.txt
Django==5.2.10
djangorestframework==3.15.2
mysqlclient==2.2.7
psycopg2-binary==2.9.9
gunicorn==23.0.0
whitenoise==6.6.0
Pillow==11.1.0
stripe==8.0.0
argon2-cffi==23.1.0
python-dotenv==1.0.0
